%% Generated from roxygen comments by tools/rd-from-roxygen.py
%% Do not edit by hand
\name{cl_bands}
\alias{cl_bands}
\title{Band names present in a scene}
\usage{
cl_bands(x)
}
\arguments{
  \item{x}{A \code{cl_scene}.}
}
\description{
Band names present in a scene
}
\value{
Character vector of standardised band names.
}
